requirements.txt
# requirements.txt
# Lista de librerías necesarias

# Procesamiento de PDFs
pdfplumber==0.11.0
camelot-py==0.11.0

# Análisis de datos
pandas==2.2.0
numpy==1.26.0

# Generación de reportes
python-docx==1.1.0
markdown==3.5.2

# Utilidades
loguru==0.7.2
python-dotenv==1.0.1
pydantic==2.5.0

# Desarrollo (opcional)
pytest==8.0.0
